+++ Adapter/IndexManager.php
@@ -126,6 +126,7 @@ class IndexManager extends IndexManagerCore
             'crossell_products_ids' => ['name' => 'crossell_products_ids', 'type' => 'string[]', 'optional' => true, 'index' => false],
             'final_price' => ['name' => 'final_price', 'type' => 'float', 'optional' => false, 'index' => true, 'sort' => true],
             'currency' => ['name' => 'currency', 'type' => 'string', 'optional' => false, 'index' => false, 'sort' => false],
+            'salable_qty' => ['name' => 'salable_qty', 'type' => 'int32', 'optional' => false, 'index' => true, 'sort' => true],
         ];
     }
 }